Add interpolation of row statements to tables
This changes the subtable mechanic for tables such that row descriptions
may include statements bracketed by '{{' and '}}'; if the statement is
parsable as an expression it is evaluated and the result injected,
otherwise the statement is looked up as a name of a table and if found
the table is recursively evaluated.

This also adds a .gitignore file for python compiled objects.