Added Template class for templated strings
This adds the Template class for evaluating templated strings with
embedded tables and rolls. This replaces the templating of the
dnd.table.Row class.

Additionally, there is a new `gems.json` table in `./tables` which
includes a selection of gems for use with treasure. Gems are split into
tables by their value, and include 1-row tables for extra formatting.